fix(workflow): port lane-ownership bars to workflow-compiler.mjs
Addresses tt-a1i re-review after rebase onto upstream/main (feat #146):

- render-workflow.mjs is now a thin compiler shim; all rendering lives
  in workflow-compiler.mjs. Ported renderLaneOwnershipBars() and
  renderPhaseBand() into workflow-compiler.mjs using the compiler's
  laneTop()/laneHeight() helpers so variable-height lanes (schema v2)
  are handled correctly.

- Lane bars use laneHeight(index) rather than the fixed layout.laneH,
  matching the compiler's own lane rendering.

- render-workflow.mjs replaced with the upstream compiler-shim version;
  no more stale old-renderer body in the file.

- Golden examples re-rendered via the compiler path.
- Lane-only fixture golden refreshed.
- test/golden.mjs synced to upstream baseline + lane-only block retained.
