+++ archify/test/fixtures/lane-ownership-no-phases.workflow.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta name="generator" content="archify 2.16.0-dev.0">
+  <meta name="generator" content="archify 2.16.0">
   <title>Product development Diagram</title>
   <script>
     // Resolve the theme before first paint so light-preference users don't
